services/ai_symptom/requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
openai
streamlit
python-dotenv